[refactor] user/fourthTab redux 리팩토링 완료
