+++ client/src/Navigator/MainTab/Customer/bidTabScreen.js
