+++ client/src/Navigator/MainTab/customerTabStack.js
@@ -2,11 +2,11 @@ import React from 'react';
 import Ionicons from 'react-native-vector-icons/Ionicons';
 import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
 
-import BidTabScreen from './User/bidTabScreen';
-import SearchTabScreen from './User/searchTabScreen';
-import HistoryTabScreen from './User/historyTabScreen';
-import MyPageTabScreen from './User/myPageTabScreen';
-import ProposalTabScreen from './User/proposalTabScreen';
+import BidTabScreen from './Customer/bidTabScreen';
+import SearchTabScreen from './Customer/searchTabScreen';
+import HistoryTabScreen from './Customer/historyTabScreen';
+import MyPageTabScreen from './Customer/myPageTabScreen';
+import ProposalTabScreen from './Customer/proposalTabScreen';
 
 const Tab = createBottomTabNavigator();
