+++ client/src/Screen/04_FourthTab/Customer/TabScreen/introStyleScrapScreen.js
