+++ client/src/Screen/04_FourthTab/Customer/historyMainScreen.js
@@ -1,8 +1,8 @@
 import React, { useState, useEffect } from 'react';
 import { createMaterialTopTabNavigator } from '@react-navigation/material-top-tabs';
 
-import StyleScrapScreen from './TabScreen/styleScrapScreen';
 import MatchingHistoryScreen from './TabScreen/matchingHistoryScreen';
+import StyleScrapScreen from './TabScreen/styleScrapScreen';
 
 const Tab = createMaterialTopTabNavigator();
 
@@ -10,7 +10,7 @@ function HistoryMainScreen({ navigation }) {
   return (
     <Tab.Navigator
       swipeEnabled={false}
-      initialRouteName="MyScrap"
+      initialRouteName="MatchingHistory"
       tabBarOptions={{
         activeTintColor: 'black',
         inactiveTintColor: '#DADADA',
@@ -26,11 +26,11 @@ function HistoryMainScreen({ navigation }) {
           borderColor: 'black',
         },
       }}>
-      <Tab.Screen name="MyScrap" options={{ title: '내 스크랩' }}>
-        {() => <StyleScrapScreen navigation={navigation} />}
-      </Tab.Screen>
       <Tab.Screen name="MatchingHistory" options={{ title: '매칭내역' }}>
         {() => <MatchingHistoryScreen />}
+      </Tab.Screen>
+      <Tab.Screen name="MyScrap" options={{ title: '내 스크랩' }}>
+        {() => <StyleScrapScreen navigation={navigation} />}
       </Tab.Screen>
     </Tab.Navigator>
   );
